# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Battle Rap API
  version: 1.0.0
  description: >
    REST API for managing online rap battles, tournaments, submissions, judging,
    and administration workflows.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/tournaments:
    get:
      tags: [Public]
      summary: List tournaments
      operationId: listTournaments
      parameters:
        - in: query
          name: status
          schema:
            type: string
          description: Filter by tournament status.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: Page number (1-indexed).
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Page size.
      responses:
        '200':
          description: Paginated tournament list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentListResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/artists:
    get:
      tags: [Public]
      summary: List public participants
      operationId: listPublicParticipants
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: Page number (1-indexed).
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Page size.
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive search across display name and full name.
        - in: query
          name: role
          schema:
            type: string
            enum: [artist, judge]
          description: Restrict listing to participants with the specified public role.
        - in: query
          name: sort
          schema:
            type: string
            enum: [joined_at, wins, rating]
          description: |
            Sorting key: `joined_at` (newest first), `wins` (descending total wins), `rating` (highest average score first).
      responses:
        '200':
          description: Paginated public participant list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicParticipantsResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/tournaments/{id}:
    get:
      tags: [Public]
      summary: Get tournament with rounds
      operationId: getTournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tournament details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament:
                    $ref: '#/components/schemas/Tournament'
                  rounds:
                    type: array
                    items:
                      $ref: '#/components/schemas/Round'
                required:
                  - tournament
                  - rounds
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/rounds/{id}:
    get:
      tags: [Public]
      summary: Get round with matches
      operationId: getRound
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Round with associated matches.
          content:
            application/json:
              schema:
                type: object
                properties:
                  round:
                    $ref: '#/components/schemas/Round'
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/MatchSummary'
                required:
                  - round
                  - matches
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/rounds/{id}/battles:
    get:
      tags: [Public]
      summary: List matches in a round with conditional winners
      operationId: listRoundBattles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Round and matches.
          content:
            application/json:
              schema:
                type: object
                properties:
                  round:
                    $ref: '#/components/schemas/Round'
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/MatchWithMaskedWinner'
                required:
                  - round
                  - matches
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/battles/{id}:
    get:
      tags: [Public]
      summary: Get battle summary
      operationId: getBattle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Battle details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Battle'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/battles/{id}/tracks:
    get:
      tags: [Public]
      summary: List published match tracks
      operationId: listBattleTracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tracks for the battle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BattleTracksResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/leaderboards:
    get:
      tags: [Public]
      summary: Get tournament leaderboard
      operationId: getLeaderboard
      parameters:
        - in: query
          name: tournament
          required: true
          schema:
            type: string
            format: uuid
          description: Tournament identifier.
      responses:
        '200':
          description: Leaderboard entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by the API. Required for authenticated endpoints.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        Refresh token cookie issued during login/registration.
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        status:
          type: integer
        field_errors:
          type: object
          additionalProperties:
            type: string
      example:
        type: https://httpstatuses.com/422
        title: Payload validation failed.
        detail: Payload validation failed.
        status: 422
        field_errors:
          email: Invalid email.
    Tournament:
      type: object
      required:
        - id
        - title
        - status
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          description: Tournament lifecycle status.
        registration_open_at:
          type: string
          format: date-time
          nullable: true
        submission_deadline_at:
          type: string
          format: date-time
          nullable: true
        judging_deadline_at:
          type: string
          format: date-time
          nullable: true
        public_at:
          type: string
          format: date-time
          nullable: true
    TournamentListResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tournament'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    Round:
      type: object
      required:
        - id
        - tournament_id
        - kind
        - number
        - scoring
        - status
      properties:
        id:
          type: string
          format: uuid
        tournament_id:
          type: string
          format: uuid
        kind:
          type: string
          description: Round type (e.g. qualification, duel).
        number:
          type: integer
        scoring:
          type: string
          enum: [pass_fail, points, rubric]
        status:
          type: string
          description: Current round state.
        starts_at:
          type: string
          format: date-time
          nullable: true
        submission_deadline_at:
          type: string
          format: date-time
          nullable: true
        judging_deadline_at:
          type: string
          format: date-time
          nullable: true
        strategy:
          type: string
          nullable: true
    MatchSummary:
      type: object
      required:
        - id
        - round_id
        - status
      properties:
        id:
          type: string
          format: uuid
        round_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        ends_at:
          type: string
          format: date-time
          nullable: true
        winner_match_track_id:
          type: string
          format: uuid
          nullable: true
    MatchWithMaskedWinner:
      allOf:
        - $ref: '#/components/schemas/MatchSummary'
    Battle:
      type: object
      required:
        - id
        - round_id
        - status
      properties:
        id:
          type: string
          format: uuid
        round_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        ends_at:
          type: string
          format: date-time
          nullable: true
        winner_match_track_id:
          type: string
          format: uuid
          nullable: true
    BattleTracksResponse:
      type: object
      required:
        - match_id
        - tracks
      properties:
        match_id:
          type: string
          format: uuid
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/MatchTrack'
    MatchTrack:
      type: object
      required:
        - id
        - participant_id
        - audio_url
        - mime
        - submitted_at
      properties:
        id:
          type: string
          format: uuid
        participant_id:
          type: string
          format: uuid
        audio_url:
          type: string
          format: uri
        mime:
          type: string
        duration_sec:
          type: number
          format: float
          nullable: true
        submitted_at:
          type: string
          format: date-time
        avg_total:
          type: number
          format: float
          nullable: true
    LeaderboardResponse:
      type: object
      required:
        - tournament_id
        - entries
      properties:
        tournament_id:
          type: string
          format: uuid
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
    LeaderboardEntry:
      type: object
      required:
        - tournament_id
        - participant_id
        - wins
      properties:
        tournament_id:
          type: string
          format: uuid
        participant_id:
          type: string
          format: uuid
        wins:
          type: integer
          minimum: 0
...
    ProfileAvatar:
      type: object
      required:
        - key
        - url
      properties:
        key:
          type: string
          description: Internal storage key for the avatar asset.
        url:
          type: string
          format: uri
          description: CDN URL for the avatar asset.
...
    PublicParticipant:
      type: object
      required:
        - id
        - display_name
        - roles
        - joined_at
        - total_wins
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        roles:
          type: array
          items:
            type: string
        city:
          type: string
          nullable: true
        joined_at:
          type: string
          format: date-time
        avatar:
          $ref: '#/components/schemas/ProfileAvatar'
          nullable: true
        avg_total_score:
          type: number
          format: float
          nullable: true
          description: Average judge score across published tracks (0-100).
        total_wins:
          type: integer
          minimum: 0
          description: Total number of recorded match wins.
    PublicParticipantsResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PublicParticipant'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...